refactor: 카테고리 및 서브 카테고리 응답 형식 description으로 변경 (#55)
