+++ src/main/java/modelly/modelly_be/domain/reservation/dto/internal/ModelReservationRow.java
@@ -1,21 +1,24 @@
-package modelly.modelly_be.domain.reservation.dto.common;
+package modelly.modelly_be.domain.reservation.dto.internal;
 
+import modelly.modelly_be.global.entity.Category;
 import modelly.modelly_be.domain.reservation.entity.enums.ReservationStatus;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public record DesignerReservationRow(
+public record ModelReservationRow(
         Long reservationId,
         LocalDate date,
         LocalTime startTime,
         LocalTime endTime,
         ReservationStatus status,
+        Category category,
 
         Long recruitmentId,
         String recruitmentTitle,
 
-        Long modelId,
-        Long modelUserId,
-        String modelName
+        Long designerUserId,
+        Long designerId,
+        String designerNickname,
+        String shop
 ) {}
