+++ src/main/java/modelly/modelly_be/domain/reservation/repository/ReservationRepositoryImpl.java
@@ -6,9 +6,9 @@ import com.querydsl.core.types.dsl.EnumPath;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
-import modelly.modelly_be.domain.reservation.dto.common.DesignerReservationRow;
+import modelly.modelly_be.domain.reservation.dto.internal.DesignerReservationRow;
 import modelly.modelly_be.domain.reservation.entity.enums.ReservationListType;
-import modelly.modelly_be.domain.reservation.dto.common.ModelReservationRow;
+import modelly.modelly_be.domain.reservation.dto.internal.ModelReservationRow;
 import modelly.modelly_be.domain.reservation.entity.enums.ReservationStatus;
 import modelly.modelly_be.global.entity.Category;
 import modelly.modelly_be.global.entity.SubCategory;
